remove forwarding server, but keep it around for now
